+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added support for templates
 - Added types of config
 - Added support for `Header` and `SectionNav` component
+- Added automatic redirect to first chapter for `/`
+- Added support for Markdown tasklists
+- Added support for Markdown (arbitrary) file inclusions
 
 ## 0.12.6 (November 5, 2021)